kalyanvarma-hiring-challenge-a7bcc12668ba/app/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
environs==11.0.0
Faker==25.8.0
marshmallow==3.21.3
packaging==24.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
